requirements.txt
flask
werkzeug
pdfminer.six
docx2txt
